+++ packages/widgetbook/lib/src/args/bool_arg.dart
@@ -1,6 +1,5 @@
 import '../core/arg.dart';
 import '../fields/fields.dart';
-import '../routing/routing.dart';
 
 class BoolArg extends Arg<bool> {
   const BoolArg(
@@ -15,16 +14,6 @@ class BoolArg extends Arg<bool> {
       initialValue: value,
     ),
   ];
-
-  @override
-  bool valueFromQueryGroup(QueryGroup group) {
-    return valueOf('value', group)!;
-  }
-
-  @override
-  QueryGroup valueToQueryGroup(bool value) {
-    return {'value': paramOf('value', value)};
-  }
 
   @override
   BoolArg init({