+++ sandbox/lib/[sam]/types_table.stories.dart
@@ -81,14 +81,6 @@ class PersonArg extends Arg<Person> {
   }
 
   @override
-  QueryGroup valueToQueryGroup(Person value) {
-    return {
-      'name': paramOf('name', value.name),
-      'age': paramOf('age', value.age),
-    };
-  }
-
-  @override
   PersonArg init({
     required String name,
   }) {